HIBERNATE 01

+++ main/java/hibernate/lesson8/dao/HotelDAO.java
@@ -1,5 +1,6 @@
-package hibernate.lesson7.dao;
-import hibernate.lesson7.models.Hotel;
+package hibernate.lesson8.dao;
+
+import hibernate.lesson8.models.Hotel;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -10,18 +11,16 @@ public class HotelDAO {
 
 
     private static SessionFactory sessionFactory;
-
-
     public Hotel save(Hotel hotel) {
         Transaction tr = null;
         try (Session session = createSessionFactory().openSession()) {
             tr = session.getTransaction();
             tr.begin();
             session.save(hotel);
-            long tempId = (long) session.save(hotel);
+          //  long tempId = (long) session.save(hotel);
             tr.commit();
             System.out.println("Done save");
-            hotel.setId(tempId);
+          //  hotel.setId(tempId);
             return hotel;
         } catch (HibernateException e) {
             System.err.println("Something went wrong during save");
@@ -54,17 +53,65 @@ public class HotelDAO {
 
     }
 
-    public Hotel delete(Hotel hotel) {
+    public void delete(Long id) {
         Transaction tr = null;
+        Hotel hotel;
         try (Session session = createSessionFactory().openSession()) {
             tr = session.getTransaction();
             tr.begin();
-            session.delete(hotel);
+            hotel = (Hotel) session.load(Hotel.class, new Long(id));
+            if(hotel!=null) {
+                session.delete(hotel);
+                System.out.println("Done delete");
+            }
             tr.commit();
-            System.out.println("Done delete");
-            return hotel;
+
         } catch (HibernateException e) {
             System.err.println("Something went wrong during delete");
+            e.printStackTrace();
+            if (tr != null) {
+                tr.rollback();
+            }
+            throw e;
+        }
+    }
+
+    public Hotel findById(Long id) {
+        Transaction tr = null;
+        Hotel hotel;
+        try (Session session = createSessionFactory().openSession()) {
+            tr = session.getTransaction();
+            tr.begin();
+            hotel = (Hotel) session.load(Hotel.class, new Long(id));
+            if(hotel!=null) {
+                System.out.println("Done findById");
+            }
+            tr.commit();
+            return hotel;
+        } catch (HibernateException e) {
+            System.err.println("Something went wrong during findById");
+            e.printStackTrace();
+            if (tr != null) {
+                tr.rollback();
+            }
+            throw e;
+        }
+    }
+
+    public Hotel findByName(String name) {
+        Transaction tr = null;
+        Hotel hotel;
+        try (Session session = createSessionFactory().openSession()) {
+            tr = session.getTransaction();
+            tr.begin();
+            hotel = (Hotel) session.load(Hotel.class, new String(name));
+            if(hotel!=null) {
+                System.out.println("Done findByName");
+            }
+            tr.commit();
+            return hotel;
+        } catch (HibernateException e) {
+            System.err.println("Something went wrong during findByName");
             e.printStackTrace();
             if (tr != null) {
                 tr.rollback();
